+++ spec/dummy/client/entries/app-kitten.js
@@ -65,6 +65,7 @@ import {
 } from 'kitten/karl/cards/similar-projects-card'
 import { KarlProjectSimilarLoader }
   from 'kitten/karl/cards/project-similar-card'
+import { KarlVerticalCardWithAction } from 'kitten/karl/cards/vertical-card-with-action'
 
 // Colors
 import { KarlColors } from 'kitten/karl/colors/karl-colors'
@@ -194,6 +195,9 @@ import { PhoneIllustration }
 // Images
 import { KarlImageCropper } from 'kitten/karl/images/image-cropper'
 import { KarlImageWithCaption } from 'kitten/karl/images/image-with-caption'
+
+// Media
+import { EmbedPlayer } from 'kitten/components/media/embed-player'
 
 // Layout
 import { Separator } from 'kitten/components/layout/separator'
@@ -382,6 +386,7 @@ ReactOnRails.register({
   KarlSimilarProjectsCard,
   KarlSimilarProjectsWithInfosCard,
   KarlProjectSimilarLoader,
+  KarlVerticalCardWithAction,
 
   // Dev
   DevGrid,
@@ -491,6 +496,9 @@ ReactOnRails.register({
   // Images
   KarlImageCropper,
   KarlImageWithCaption,
+
+  //Media
+  EmbedPlayer,
 
   // Layout
   Separator,