+++ CHANGELOG.md
@@ -5,7 +5,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add props on `ProjectCard` and `CrowdfundingCard` components.
-- Fix: Fix `button` and `image` position on `ContributionCard` component.
+
+## [23.3.0] - 2018-09-18
+
+Features:
+- Update style on `SOrLess` version on `ContributionCard` component.
+- Create `VerticalCardWithAction` component.
+- Create `EmbedPlayer` component.
+
+Fix:
+- Fix `button` and `image` position on `ContributionCard` component.
 
 ## [23.2.1] - 2018-09-04